refactor: decompose environment controller reconciliation
Amp-Thread-ID: https://ampcode.com/threads/T-019fa4c3-3e7c-718f-843c-301de9bb0b5d
