+++ internal/controllers/environment_controller.go
@@ -16,7 +16,6 @@ import (
 	"fmt"
 	"math"
 	"math/big"
-	"net"
 	"strconv"
 	"strings"
 	"time"
@@ -34,17 +33,10 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/util/retry"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	platformv1alpha1 "github.com/Chris-Cullins/swe-platform/api/v1alpha1"
-	"github.com/Chris-Cullins/swe-platform/internal/lifecycle"
 	"github.com/Chris-Cullins/swe-platform/internal/tenancy"
 	sandboxdauth "github.com/Chris-Cullins/swe-platform/sandboxd/auth"
 )
@@ -162,479 +154,6 @@ type EnvironmentReconciler struct {
 // +kubebuilder:rbac:groups="",resources=secrets,verbs=get;create;update;delete
 // +kubebuilder:rbac:groups=networking.k8s.io,resources=networkpolicies,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=node.k8s.io,resources=runtimeclasses,verbs=get;list;watch
-
-// Reconcile drives an Environment toward its desired state:
-// pod + PVC present when active, pod deleted (PVC retained) when paused.
-func (r *EnvironmentReconciler) Reconcile(ctx context.Context, req ctrl.Request) (result ctrl.Result, err error) {
-	defer func() {
-		if stderrors.Is(err, errEnvironmentIncarnationChanged) {
-			result = ctrl.Result{}
-			err = nil
-		}
-	}()
-
-	var env platformv1alpha1.Environment
-	if err := r.apiReader().Get(ctx, req.NamespacedName, &env); err != nil {
-		return ctrl.Result{}, client.IgnoreNotFound(err)
-	}
-	ctx, namespaceClaim, err := r.Scope.Begin(ctx, env.Namespace, tenancy.LifecycleActive, tenancy.LifecycleFencing)
-	if err != nil {
-		if stderrors.Is(err, tenancy.ErrOutOfScope) {
-			return ctrl.Result{}, nil
-		}
-		return ctrl.Result{}, err
-	}
-	if namespaceClaim.Lifecycle == tenancy.LifecycleFencing {
-		if namespaceClaim.Operation != tenancy.OperationOffboarding {
-			return ctrl.Result{}, nil
-		}
-		if env.DeletionTimestamp.IsZero() && (env.Spec.Lifecycle.Hold == nil || !env.Spec.Lifecycle.Hold.Enabled) {
-			before := env.DeepCopy()
-			revision := int64(1)
-			if env.Spec.Lifecycle.Hold != nil {
-				revision = env.Spec.Lifecycle.Hold.Revision + 1
-			}
-			env.Spec.Lifecycle.Hold = &platformv1alpha1.EnvironmentHoldPolicy{Enabled: true, Revision: revision}
-			if err := r.Patch(ctx, &env, client.MergeFromWithOptions(before, client.MergeFromWithOptimisticLock{})); err != nil {
-				return ctrl.Result{}, fmt.Errorf("fence Environment for Project offboarding: %w", err)
-			}
-			return ctrl.Result{Requeue: true}, nil
-		}
-	}
-
-	if !env.DeletionTimestamp.IsZero() {
-		return r.reconcileDeleting(ctx, &env)
-	}
-	if !controllerutil.ContainsFinalizer(&env, environmentFinalizer) {
-		controllerutil.AddFinalizer(&env, environmentFinalizer)
-		if err := r.Update(ctx, &env); err != nil {
-			return ctrl.Result{}, err
-		}
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if env.Spec.Paused {
-		before := env.DeepCopy()
-		if env.Spec.Lifecycle.Hold == nil {
-			env.Spec.Lifecycle.Hold = &platformv1alpha1.EnvironmentHoldPolicy{Enabled: true, Revision: 1}
-		} else if !env.Spec.Lifecycle.Hold.Enabled {
-			env.Spec.Lifecycle.Hold.Enabled = true
-			env.Spec.Lifecycle.Hold.Revision++
-		}
-		env.Spec.Paused = false
-		if err := r.Patch(ctx, &env, client.MergeFromWithOptions(before, client.MergeFromWithOptimisticLock{})); err != nil {
-			return ctrl.Result{}, fmt.Errorf("migrate deprecated paused intent: %w", err)
-		}
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if request := env.Spec.Lifecycle.Suspend; request != nil &&
-		(env.Status.Lifecycle.PendingSuspendRequestID != request.ID || env.Status.Lifecycle.LastSuspendRequestSequence != request.Sequence) &&
-		(request.Sequence <= env.Status.Lifecycle.LastSuspendRequestSequence || request.ID == env.Status.Lifecycle.LastSuspendRequestID || !validLifecycleRequest(&env, &request.EnvironmentLifecycleRequest)) {
-		before := env.DeepCopy()
-		env.Spec.Lifecycle.Suspend = nil
-		if err := r.Patch(ctx, &env, client.MergeFromWithOptions(before, client.MergeFromWithOptimisticLock{})); err != nil {
-			return ctrl.Result{}, fmt.Errorf("discard replayed suspension request: %w", err)
-		}
-		return ctrl.Result{Requeue: true}, nil
-	}
-	changed, err := r.reconcileLifecycleIntent(ctx, &env)
-	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("reconcile lifecycle intent: %w", err)
-	}
-	if changed {
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if invalidProvisioningFenceStarted(&env) {
-		if result, handled, err := r.reconcileInvalidProvisioningFence(ctx, &env); handled || err != nil {
-			return result, err
-		}
-	}
-	project, projectErr := r.resolveEnvironmentProject(ctx, &env)
-	if project != nil && len(project.Spec.EgressAllowlist) != 0 {
-		return r.reconcileInvalidProvisioningConfiguration(ctx, &env, unsupportedEgressAllowlistMessage(project.Name))
-	}
-	// Fencing must not depend on a still-readable template or successful setup.
-	// Cancellation/finalization can therefore stop an execution domain even after
-	// its template or Project was deleted or provisioning became permanently broken.
-	if env.Status.Lifecycle.Suspended {
-		result, err := r.reconcilePaused(ctx, &env)
-		if err != nil {
-			return ctrl.Result{}, r.fail(ctx, &env, fmt.Errorf("pause environment: %w", err))
-		}
-		return result, nil
-	}
-	if projectErr != nil {
-		var terminal *terminalEnvironmentError
-		if stderrors.As(projectErr, &terminal) {
-			return r.reconcileInvalidProvisioningConfiguration(ctx, &env, projectErr.Error())
-		}
-		return ctrl.Result{}, r.fail(ctx, &env, projectErr)
-	}
-	if err := validateEnvironmentProject(project); err != nil {
-		return ctrl.Result{}, r.fail(ctx, &env, err)
-	}
-	if strings.TrimSpace(env.Spec.TemplateRef) == "" {
-		return r.reconcileInvalidProvisioningConfiguration(ctx, &env, "environment templateRef must not be blank")
-	}
-	var tmpl platformv1alpha1.EnvironmentTemplate
-	if err := r.Get(ctx, types.NamespacedName{Namespace: env.Namespace, Name: env.Spec.TemplateRef}, &tmpl); err != nil {
-		wrapped := fmt.Errorf("get template %q: %w", env.Spec.TemplateRef, err)
-		if errors.IsNotFound(err) {
-			return r.reconcileInvalidProvisioningConfiguration(ctx, &env, wrapped.Error())
-		}
-		return ctrl.Result{}, r.fail(ctx, &env, wrapped)
-	}
-	if tenancy.IsCatalogSource(&tmpl) {
-		return r.reconcileInvalidProvisioningConfiguration(ctx, &env, fmt.Sprintf("environment template %q is an inert installation catalog source", tmpl.Name))
-	}
-	if r.Scope != nil && r.Scope.Verifier != nil {
-		if err := tenancy.ValidateManagedTemplate(&tmpl, r.Scope.Verifier.Installation, namespaceClaim); err != nil {
-			return r.reconcileInvalidProvisioningConfiguration(ctx, &env, err.Error())
-		}
-	}
-	if result, handled, err := r.reconcilePendingPodRecovery(ctx, &env); handled || err != nil {
-		return result, err
-	}
-	backend := platformv1alpha1.EffectiveEnvironmentBackend(&env, &tmpl)
-	if backend != platformv1alpha1.EnvironmentBackendPod {
-		return r.reconcileUnsupportedBackend(ctx, &env, backend)
-	}
-	runtimeClassUID := types.UID("")
-	if tmpl.Spec.RuntimeClass != "" {
-		var runtimeClass nodev1.RuntimeClass
-		if err := r.Get(ctx, types.NamespacedName{Name: tmpl.Spec.RuntimeClass}, &runtimeClass); err != nil {
-			wrapped := fmt.Errorf("get RuntimeClass %q required by template %q: %w", tmpl.Spec.RuntimeClass, tmpl.Name, err)
-			if errors.IsNotFound(err) {
-				return r.reconcileInvalidProvisioningConfiguration(ctx, &env, wrapped.Error())
-			}
-			return ctrl.Result{}, r.fail(ctx, &env, wrapped)
-		}
-		runtimeClassUID = runtimeClass.UID
-		var pod corev1.Pod
-		if err := r.apiReader().Get(ctx, types.NamespacedName{Namespace: env.Namespace, Name: envPodName(&env)}, &pod); err == nil {
-			if exactControllerOwner(&pod, platformv1alpha1.GroupVersion.String(), "Environment", env.Name, env.UID) &&
-				pod.Annotations[runtimeClassUIDAnnotation] != string(runtimeClassUID) {
-				message := fmt.Sprintf("environment pod RuntimeClass %q incarnation does not match the current RuntimeClass; execution must be replaced", tmpl.Spec.RuntimeClass)
-				return r.reconcileInvalidProvisioningConfiguration(ctx, &env, message)
-			}
-		} else if !errors.IsNotFound(err) {
-			return ctrl.Result{}, r.fail(ctx, &env, fmt.Errorf("get environment pod before RuntimeClass validation: %w", err))
-		}
-	}
-
-	pvcReady, err := r.ensureWorkspacePVC(ctx, &env, &tmpl)
-	if err != nil {
-		return ctrl.Result{}, r.fail(ctx, &env, fmt.Errorf("ensure workspace PVC: %w", err))
-	}
-	if !pvcReady {
-		return ctrl.Result{Requeue: true}, nil
-	}
-	policyReady, err := r.ensureSandboxdNetworkPolicy(ctx, &env)
-	if err != nil {
-		return ctrl.Result{}, r.fail(ctx, &env, fmt.Errorf("ensure sandboxd network policy: %w", err))
-	}
-	if !policyReady {
-		return ctrl.Result{Requeue: true}, nil
-	}
-
-	if env.Status.Phase == platformv1alpha1.EnvironmentPhasePaused {
-		now := metav1.Now()
-		env.Status.LastActiveAt = &now
-		return ctrl.Result{Requeue: true}, r.setPhase(ctx, &env, platformv1alpha1.EnvironmentPhaseResuming, "", "")
-	}
-
-	pod, err := r.ensurePodForProject(ctx, &env, &tmpl, project, runtimeClassUID)
-	if stderrors.Is(err, errPodReplacing) {
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if err != nil {
-		return ctrl.Result{}, r.fail(ctx, &env, fmt.Errorf("ensure pod: %w", err))
-	}
-	if pod == nil {
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if pod.Status.Phase == corev1.PodFailed || pod.Status.Phase == corev1.PodSucceeded {
-		return r.reconcileTerminalPod(ctx, &env, pod)
-	}
-
-	if err := r.syncStatus(ctx, &env, pod); err != nil {
-		return ctrl.Result{}, err
-	}
-	if pod.Status.Phase != corev1.PodRunning {
-		return ctrl.Result{}, nil
-	}
-	result, err = r.reconcileIdle(ctx, &env, &tmpl)
-	if errors.IsConflict(err) {
-		// A concurrent claim or activity update won the optimistic pause race.
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if err != nil {
-		return ctrl.Result{}, r.fail(ctx, &env, fmt.Errorf("reconcile idle policy: %w", err))
-	}
-	return result, nil
-}
-
-// reconcileLifecycleIntent is the sole owner of lifecycle transitions. Other
-// components publish fenced intents in spec or bounded metadata slots; this
-// method consumes each valid request once and decides whether execution is suspended.
-func (r *EnvironmentReconciler) reconcileLifecycleIntent(ctx context.Context, env *platformv1alpha1.Environment) (bool, error) {
-	activityRequests := lifecycle.ActivityRequests(env)
-	if env.Spec.Lifecycle.Hold == nil && env.Spec.Lifecycle.Wake == nil && env.Spec.Lifecycle.Suspend == nil && len(activityRequests) == 0 &&
-		!env.Status.Lifecycle.Suspended && env.Status.Lifecycle.Epoch == 0 && env.Status.Lifecycle.ObservedHoldPolicyRevision == 0 &&
-		env.Status.Lifecycle.LastWakeRequestID == "" && env.Status.Lifecycle.LastSuspendRequestID == "" && env.Status.Lifecycle.LastSuspendRequestSequence == 0 &&
-		env.Status.Lifecycle.PendingSuspendRequestID == "" && len(env.Status.Lifecycle.ActivityReceipts) == 0 {
-		return false, nil
-	}
-	policyRevision := int64(0)
-	held := false
-	if env.Spec.Lifecycle.Hold != nil {
-		policyRevision = env.Spec.Lifecycle.Hold.Revision
-		held = env.Spec.Lifecycle.Hold.Enabled
-	}
-	valid := func(request *platformv1alpha1.EnvironmentLifecycleRequest) bool {
-		return request != nil && request.EnvironmentUID == env.UID && request.HoldPolicyRevision == policyRevision
-	}
-
-	before := env.Status.DeepCopy()
-	wasSuspended := env.Status.Lifecycle.Suspended
-	previousReason := env.Status.Lifecycle.SuspensionReason
-	lifecycleStatus := &env.Status.Lifecycle
-	lifecycleStatus.ObservedHoldPolicyRevision = policyRevision
-	now := metav1.NewTime(r.now())
-	for i := range activityRequests {
-		request := &activityRequests[i]
-		if !valid(&request.EnvironmentLifecycleRequest) || request.ExecutionGeneration != env.Status.ExecutionGeneration ||
-			activityReceipt(lifecycleStatus.ActivityReceipts, request.Source, request.ExecutionGeneration) == request.ID {
-			continue
-		}
-		setActivityReceipt(&lifecycleStatus.ActivityReceipts, request.Source, request.ExecutionGeneration, request.ID)
-		if env.Status.LastActiveAt == nil || now.After(env.Status.LastActiveAt.Time) {
-			env.Status.LastActiveAt = &now
-		}
-	}
-
-	suspend := env.Spec.Lifecycle.Suspend
-	suspendValid := suspend != nil && valid(&suspend.EnvironmentLifecycleRequest)
-	suspendInFlight := lifecycleStatus.PendingSuspendRequestID != ""
-	suspendNew := suspendValid && !suspendInFlight && suspend.Sequence > lifecycleStatus.LastSuspendRequestSequence && suspend.ID != lifecycleStatus.LastSuspendRequestID
-	if suspendNew {
-		lifecycleStatus.LastSuspendRequestID = suspend.ID
-		lifecycleStatus.LastSuspendRequestSequence = suspend.Sequence
-		lifecycleStatus.PendingSuspendRequestID = suspend.ID
-	}
-	wake := env.Spec.Lifecycle.Wake
-	wakeNew := wake != nil && valid(&wake.EnvironmentLifecycleRequest) && wake.ID != lifecycleStatus.LastWakeRequestID
-	wakeSuspensionReason := lifecycleStatus.SuspensionReason
-	switch {
-	case held:
-		wakeSuspensionReason = platformv1alpha1.EnvironmentSuspensionReasonHold
-	case suspendNew || suspendInFlight:
-		wakeSuspensionReason = platformv1alpha1.EnvironmentSuspensionReasonRequested
-	}
-	wakeReasonMatches := false
-	if wake != nil {
-		expectedReason := wake.ExpectedSuspensionReason
-		if expectedReason == "" {
-			expectedReason = platformv1alpha1.EnvironmentSuspensionReasonIdle
-		}
-		wakeReasonMatches = expectedReason == wakeSuspensionReason
-	}
-	backendFencePending := lifecycleStatus.Suspended && env.Status.Phase != platformv1alpha1.EnvironmentPhasePaused
-	consumeWake := wakeNew && (held || !wakeReasonMatches || !suspendNew && !suspendInFlight && !backendFencePending)
-	if consumeWake {
-		lifecycleStatus.LastWakeRequestID = wake.ID
-	}
-
-	suspended := lifecycleStatus.Suspended
-	reason := lifecycleStatus.SuspensionReason
-	requestID := lifecycleStatus.SuspensionRequestID
-	switch {
-	case held:
-		suspended = true
-		reason = platformv1alpha1.EnvironmentSuspensionReasonHold
-		requestID = ""
-	case suspendNew || suspendInFlight:
-		suspended = true
-		reason = platformv1alpha1.EnvironmentSuspensionReasonRequested
-		requestID = lifecycleStatus.PendingSuspendRequestID
-	case lifecycleStatus.Suspended && lifecycleStatus.SuspensionReason == platformv1alpha1.EnvironmentSuspensionReasonHold && !backendFencePending:
-		// Disabling a hold at a newer policy revision is itself the authorized
-		// release; no ordinary stale wake is required. The backend fence must
-		// complete first so release cannot reuse the previous execution domain.
-		suspended = false
-		reason = ""
-		requestID = ""
-	case consumeWake && wakeReasonMatches && lifecycleStatus.SuspensionReason != platformv1alpha1.EnvironmentSuspensionReasonHold:
-		suspended = false
-		reason = ""
-		requestID = ""
-	}
-	if suspended && !lifecycleStatus.Suspended {
-		lifecycleStatus.Epoch++
-	}
-	lifecycleStatus.Suspended = suspended
-	lifecycleStatus.SuspensionReason = reason
-	lifecycleStatus.SuspensionRequestID = requestID
-
-	if apiequality.Semantic.DeepEqual(*before, env.Status) {
-		return false, nil
-	}
-	if err := r.Status().Update(ctx, env); err != nil {
-		return false, err
-	}
-	if !wasSuspended && lifecycleStatus.Suspended {
-		r.Metrics.observeLifecycle("suspend", lifecycleStatus.SuspensionReason)
-	} else if wasSuspended && !lifecycleStatus.Suspended {
-		r.Metrics.observeLifecycle("resume", previousReason)
-	}
-	return true, nil
-}
-
-func validLifecycleRequest(env *platformv1alpha1.Environment, request *platformv1alpha1.EnvironmentLifecycleRequest) bool {
-	policyRevision := int64(0)
-	if env.Spec.Lifecycle.Hold != nil {
-		policyRevision = env.Spec.Lifecycle.Hold.Revision
-	}
-	return request.EnvironmentUID == env.UID && request.HoldPolicyRevision == policyRevision
-}
-
-func activityReceipt(receipts []platformv1alpha1.EnvironmentActivityReceipt, source platformv1alpha1.EnvironmentActivitySource, executionGeneration int64) string {
-	for i := range receipts {
-		if receipts[i].Source == source && receipts[i].ExecutionGeneration == executionGeneration {
-			return receipts[i].RequestID
-		}
-	}
-	return ""
-}
-
-func setActivityReceipt(receipts *[]platformv1alpha1.EnvironmentActivityReceipt, source platformv1alpha1.EnvironmentActivitySource, executionGeneration int64, requestID string) {
-	for i := range *receipts {
-		if (*receipts)[i].Source == source {
-			(*receipts)[i].RequestID = requestID
-			(*receipts)[i].ExecutionGeneration = executionGeneration
-			return
-		}
-	}
-	*receipts = append(*receipts, platformv1alpha1.EnvironmentActivityReceipt{Source: source, RequestID: requestID, ExecutionGeneration: executionGeneration})
-}
-
-// reconcilePendingPodRecovery advances persisted recovery before ensurePod can
-// create a replacement. This keeps backoff and exhaustion effective even when
-// a terminal Pod disappears without the controller deleting it.
-func (r *EnvironmentReconciler) reconcilePendingPodRecovery(ctx context.Context, env *platformv1alpha1.Environment) (ctrl.Result, bool, error) {
-	ready := apimeta.FindStatusCondition(env.Status.Conditions, platformv1alpha1.EnvironmentConditionReady)
-	if env.Status.PodRecoveryExhausted {
-		message := fmt.Sprintf("automatic recovery is exhausted after %d terminal pod replacements", env.Status.PodRecoveryAttempts)
-		if ready == nil || ready.ObservedGeneration != env.Generation || ready.Status != metav1.ConditionFalse || ready.Reason != "PodRecoveryExhausted" {
-			if err := r.setEnvironmentStatus(ctx, env, platformv1alpha1.EnvironmentPhaseFailed, env.Status.PodName, "", "PodRecoveryExhausted", message); err != nil {
-				return ctrl.Result{}, true, err
-			}
-		}
-		return ctrl.Result{}, true, nil
-	}
-	nextAttemptAt := env.Status.PodRecoveryNextAttemptAt
-	if nextAttemptAt == nil {
-		return ctrl.Result{}, false, nil
-	}
-	now := r.now()
-	if now.Before(nextAttemptAt.Time) {
-		if ready == nil || ready.ObservedGeneration != env.Generation || ready.Status != metav1.ConditionFalse || ready.Reason != "PodRecoveryPending" {
-			message := fmt.Sprintf("terminal pod recovery attempt %d of %d is scheduled for %s", env.Status.PodRecoveryAttempts+1, podRecoveryLimit, nextAttemptAt.Time.UTC().Format(time.RFC3339))
-			if err := r.setEnvironmentStatus(ctx, env, platformv1alpha1.EnvironmentPhaseCreating, "", "", "PodRecoveryPending", message); err != nil {
-				return ctrl.Result{}, true, err
-			}
-		}
-		return ctrl.Result{RequeueAfter: nextAttemptAt.Sub(now)}, true, nil
-	}
-
-	attempts := env.Status.PodRecoveryAttempts + 1
-	message := fmt.Sprintf("replacing terminal environment pod (recovery attempt %d of %d)", attempts, podRecoveryLimit)
-	if err := r.updatePodRecoveryStatus(ctx, env, func(current *platformv1alpha1.Environment) {
-		applyEnvironmentStatus(current, platformv1alpha1.EnvironmentPhaseCreating, "", "", "PodRecovering", message, env.Status.LastActiveAt)
-		current.Status.PodRecoveryAttempts = attempts
-		current.Status.PodRecoveryUID = env.Status.PodRecoveryUID
-		current.Status.PodRecoveryNextAttemptAt = nil
-		clearChildOwnershipCollision(current)
-	}); err != nil {
-		if stderrors.Is(err, errPodRecoveryChanged) {
-			return ctrl.Result{Requeue: true}, true, nil
-		}
-		return ctrl.Result{}, true, err
-	}
-	r.Metrics.observePodRecovery("attempt")
-	// Reconcile again before deleting or creating so the persisted attempt marker
-	// is always observed, including across a concurrent generation change.
-	return ctrl.Result{Requeue: true}, true, nil
-}
-
-// reconcileTerminalPod replaces an owned terminal Pod using a persisted,
-// bounded retry budget. The Pod UID fences both the delay and delete
-// so retries after controller or API failures cannot consume the budget twice
-// or delete a same-name replacement.
-func (r *EnvironmentReconciler) reconcileTerminalPod(ctx context.Context, env *platformv1alpha1.Environment, pod *corev1.Pod) (ctrl.Result, error) {
-	now := r.now()
-	attempts := env.Status.PodRecoveryAttempts
-	recoveryUID := env.Status.PodRecoveryUID
-	nextAttemptAt := env.Status.PodRecoveryNextAttemptAt
-
-	if recoveryUID != pod.UID {
-		if attempts >= podRecoveryLimit {
-			message := fmt.Sprintf("environment pod %s after %d recovery attempts; automatic recovery is exhausted", strings.ToLower(string(pod.Status.Phase)), attempts)
-			if err := r.updatePodRecoveryStatus(ctx, env, func(current *platformv1alpha1.Environment) {
-				applyEnvironmentStatus(current, platformv1alpha1.EnvironmentPhaseFailed, pod.Name, "", "PodRecoveryExhausted", message, env.Status.LastActiveAt)
-				current.Status.PodRecoveryExhausted = true
-				clearChildOwnershipCollision(current)
-			}); err != nil {
-				if stderrors.Is(err, errPodRecoveryChanged) {
-					return ctrl.Result{Requeue: true}, nil
-				}
-				return ctrl.Result{}, err
-			}
-			r.Metrics.observePodRecovery("exhausted")
-			log.FromContext(ctx).Info("environment pod recovery exhausted", "environment", env.Name, "pod", pod.Name, "attempts", attempts)
-			return ctrl.Result{}, nil
-		}
-		next := metav1.NewTime(now.Add(podRecoveryBackoff(attempts)))
-		message := fmt.Sprintf("environment pod %s; recovery attempt %d of %d is scheduled for %s", strings.ToLower(string(pod.Status.Phase)), attempts+1, podRecoveryLimit, next.Time.UTC().Format(time.RFC3339))
-		if err := r.updatePodRecoveryStatus(ctx, env, func(current *platformv1alpha1.Environment) {
-			applyEnvironmentStatus(current, platformv1alpha1.EnvironmentPhaseCreating, "", "", "PodRecoveryPending", message, env.Status.LastActiveAt)
-			current.Status.PodRecoveryAttempts = attempts
-			current.Status.PodRecoveryExhausted = false
-			current.Status.PodRecoveryUID = pod.UID
-			current.Status.PodRecoveryNextAttemptAt = &next
-			clearChildOwnershipCollision(current)
-		}); err != nil {
-			if stderrors.Is(err, errPodRecoveryChanged) {
-				return ctrl.Result{Requeue: true}, nil
-			}
-			return ctrl.Result{}, err
-		}
-		log.FromContext(ctx).Info("scheduled environment pod recovery", "environment", env.Name, "pod", pod.Name, "attempt", attempts+1, "maxAttempts", podRecoveryLimit, "nextAttemptAt", next.Time)
-		return ctrl.Result{RequeueAfter: podRecoveryBackoff(attempts)}, nil
-	}
-
-	if nextAttemptAt != nil {
-		if now.Before(nextAttemptAt.Time) {
-			return ctrl.Result{RequeueAfter: nextAttemptAt.Sub(now)}, nil
-		}
-		return ctrl.Result{Requeue: true}, nil
-	}
-	if err := r.deleteObservedChild(ctx, pod); err != nil && !errors.IsNotFound(err) {
-		return ctrl.Result{}, fmt.Errorf("delete terminal pod for recovery: %w", err)
-	}
-	log.FromContext(ctx).Info("replacing terminal environment pod", "environment", env.Name, "pod", pod.Name, "attempt", attempts, "maxAttempts", podRecoveryLimit)
-	return ctrl.Result{Requeue: true}, nil
-}
-
-func podRecoveryBackoff(attempts int32) time.Duration {
-	return podRecoveryDelay * time.Duration(1<<attempts)
-}
-
-func (r *EnvironmentReconciler) now() time.Time {
-	if r.Now != nil {
-		return r.Now()
-	}
-	return time.Now()
-}
 
 // reconcileInvalidProvisioningConfiguration withdraws the published
 // connection before fencing an execution domain whose referenced provisioning
@@ -1499,232 +1018,6 @@ func (r *EnvironmentReconciler) ensureSandboxdNetworkPolicy(ctx context.Context,
 
 // syncStatus maps Kubernetes-native pod readiness and container failure state
 // onto the Environment's generation-aware readiness contract.
-func (r *EnvironmentReconciler) syncStatus(ctx context.Context, env *platformv1alpha1.Environment, pod *corev1.Pod) error {
-	executionGeneration, ok := podExecutionGeneration(pod)
-	if !ok || executionGeneration != env.Status.ExecutionGeneration {
-		return fmt.Errorf("environment pod execution generation is not current")
-	}
-	phase, reason, message := environmentPodState(env, pod)
-	sandboxdEndpoint := ""
-	if phase == platformv1alpha1.EnvironmentPhaseReady {
-		sandboxdEndpoint = net.JoinHostPort(pod.Status.PodIP, "50051")
-		if env.Status.LastActiveAt == nil || (env.Status.Phase != platformv1alpha1.EnvironmentPhaseReady && env.Status.Phase != platformv1alpha1.EnvironmentPhaseRunning) {
-			now := metav1.Now()
-			env.Status.LastActiveAt = &now
-		}
-	}
-	return r.updateEnvironmentStatus(ctx, env, func(current *platformv1alpha1.Environment) {
-		applyEnvironmentStatus(current, phase, pod.Name, sandboxdEndpoint, reason, message, env.Status.LastActiveAt)
-		current.Status.ImageID = environmentImageID(pod)
-		if phase == platformv1alpha1.EnvironmentPhaseReady {
-			current.Status.PodRecoveryAttempts = 0
-			current.Status.PodRecoveryExhausted = false
-			current.Status.PodRecoveryUID = ""
-			current.Status.PodRecoveryNextAttemptAt = nil
-		}
-		clearChildOwnershipCollision(current)
-	})
-}
-
-func environmentImageID(pod *corev1.Pod) string {
-	for _, status := range pod.Status.ContainerStatuses {
-		if status.Name == "environment" {
-			return status.ImageID
-		}
-	}
-	return ""
-}
-
-func environmentPodState(env *platformv1alpha1.Environment, pod *corev1.Pod) (platformv1alpha1.EnvironmentPhase, string, string) {
-	resuming := podIsResume(pod) || env.Status.Phase == platformv1alpha1.EnvironmentPhaseResuming
-	for _, condition := range pod.Status.Conditions {
-		if condition.Type == corev1.PodScheduled && condition.Status == corev1.ConditionFalse && condition.Reason == corev1.PodReasonUnschedulable {
-			return platformv1alpha1.EnvironmentPhaseCreating, "Unschedulable", messageOr(condition.Message, "the scheduler cannot currently place the environment pod")
-		}
-	}
-	for _, status := range pod.Status.InitContainerStatuses {
-		if terminated := status.State.Terminated; terminated != nil && terminated.ExitCode != 0 {
-			return initContainerFailure(status.Name, terminated, resuming)
-		}
-		if status.State.Waiting != nil {
-			waiting := status.State.Waiting
-			if terminated := status.LastTerminationState.Terminated; terminated != nil && terminated.ExitCode != 0 {
-				return initContainerFailure(status.Name, terminated, resuming)
-			}
-			if imagePullFailure(waiting.Reason) {
-				return platformv1alpha1.EnvironmentPhaseFailed, "ImagePullFailed", containerStatusMessage(status.Name, waiting.Reason, waiting.Message)
-			}
-			if waiting.Reason == "CrashLoopBackOff" {
-				return platformv1alpha1.EnvironmentPhaseFailed, setupReason(resuming, "Failed"), containerStatusMessage(status.Name, waiting.Reason, waiting.Message)
-			}
-		}
-	}
-	for _, status := range pod.Status.ContainerStatuses {
-		if status.State.Waiting != nil {
-			waiting := status.State.Waiting
-			if imagePullFailure(waiting.Reason) {
-				return platformv1alpha1.EnvironmentPhaseFailed, "ImagePullFailed", containerStatusMessage(status.Name, waiting.Reason, waiting.Message)
-			}
-			if waiting.Reason == "CrashLoopBackOff" {
-				return platformv1alpha1.EnvironmentPhaseFailed, "SandboxdCrashLoopBackOff", containerStatusMessage(status.Name, waiting.Reason, waiting.Message)
-			}
-		}
-	}
-
-	switch pod.Status.Phase {
-	case corev1.PodPending:
-		if resuming {
-			return platformv1alpha1.EnvironmentPhaseResuming, "ResumeInProgress", "repository resume and sandboxd startup are in progress"
-		}
-		if len(pod.Spec.InitContainers) > 0 {
-			return platformv1alpha1.EnvironmentPhaseSetup, "SetupInProgress", "repository setup and sandboxd startup are in progress"
-		}
-		return platformv1alpha1.EnvironmentPhaseCreating, "Provisioning", "environment pod is provisioning"
-	case corev1.PodRunning:
-		if podReady(pod) && pod.Status.PodIP != "" {
-			return platformv1alpha1.EnvironmentPhaseReady, "SandboxdReady", "setup is complete and sandboxd is ready"
-		}
-		return platformv1alpha1.EnvironmentPhaseCreating, "SandboxdNotReady", "sandboxd has not passed its readiness probe"
-	case corev1.PodFailed:
-		return platformv1alpha1.EnvironmentPhaseFailed, "PodFailed", messageOr(pod.Status.Message, "environment pod failed")
-	case corev1.PodSucceeded:
-		return platformv1alpha1.EnvironmentPhaseTerminated, "PodTerminated", "environment pod terminated"
-	default:
-		return platformv1alpha1.EnvironmentPhaseCreating, "Provisioning", "environment pod is provisioning"
-	}
-}
-
-func initContainerFailure(name string, terminated *corev1.ContainerStateTerminated, resuming bool) (platformv1alpha1.EnvironmentPhase, string, string) {
-	reason := setupReason(resuming, "Failed")
-	if terminated.ExitCode == 124 || terminated.ExitCode == 137 {
-		reason = setupReason(resuming, "HookTimedOut")
-	}
-	message := fmt.Sprintf("init container %s exited with code %d", name, terminated.ExitCode)
-	if terminated.Message != "" {
-		message += ": " + terminated.Message
-	}
-	return platformv1alpha1.EnvironmentPhaseFailed, reason, message
-}
-
-func messageOr(message, fallback string) string {
-	if message != "" {
-		return message
-	}
-	return fallback
-}
-
-func podIsResume(pod *corev1.Pod) bool {
-	for _, container := range pod.Spec.InitContainers {
-		for _, variable := range container.Env {
-			if variable.Name == "SWE_RESUMING" && variable.Value == "true" {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func imagePullFailure(reason string) bool {
-	return reason == "ErrImagePull" || reason == "ImagePullBackOff" || reason == "InvalidImageName"
-}
-
-func setupReason(resuming bool, suffix string) string {
-	if resuming {
-		return "Resume" + suffix
-	}
-	return "Setup" + suffix
-}
-
-func containerStatusMessage(name, reason, message string) string {
-	if message == "" {
-		return fmt.Sprintf("container %s: %s", name, reason)
-	}
-	return fmt.Sprintf("container %s: %s: %s", name, reason, message)
-}
-
-func podReady(pod *corev1.Pod) bool {
-	for _, condition := range pod.Status.Conditions {
-		if condition.Type == corev1.PodReady {
-			return condition.Status == corev1.ConditionTrue
-		}
-	}
-	return false
-}
-
-func (r *EnvironmentReconciler) setPhase(ctx context.Context, env *platformv1alpha1.Environment, phase platformv1alpha1.EnvironmentPhase, podName, sandboxdEndpoint string) error {
-	reason, message := phaseReadiness(phase)
-	return r.setEnvironmentStatus(ctx, env, phase, podName, sandboxdEndpoint, reason, message)
-}
-
-func (r *EnvironmentReconciler) fail(ctx context.Context, env *platformv1alpha1.Environment, err error) error {
-	log.FromContext(ctx).Error(err, "reconcile failed", "environment", env.Name)
-	var collision *childOwnershipCollisionError
-	var terminal *terminalEnvironmentError
-	phase := platformv1alpha1.EnvironmentPhaseCreating
-	reason := "OperationalError"
-	if stderrors.As(err, &terminal) || errors.IsInvalid(err) || errors.IsBadRequest(err) {
-		phase = platformv1alpha1.EnvironmentPhaseFailed
-		reason = "InvalidConfiguration"
-	} else if stderrors.As(err, &collision) {
-		phase = platformv1alpha1.EnvironmentPhaseFailed
-		reason = "ResourceCollision"
-	}
-	statusErr := r.updateEnvironmentStatus(ctx, env, func(current *platformv1alpha1.Environment) {
-		applyEnvironmentStatus(current, phase, "", "", reason, err.Error(), env.Status.LastActiveAt)
-		if stderrors.As(err, &collision) {
-			apimeta.SetStatusCondition(&current.Status.Conditions, metav1.Condition{
-				Type:               "ChildOwnershipConflict",
-				Status:             metav1.ConditionTrue,
-				ObservedGeneration: current.Generation,
-				Reason:             "ResourceCollision",
-				Message:            collision.Error(),
-			})
-		}
-	})
-	if statusErr != nil {
-		return statusErr
-	}
-	if collision != nil || terminal != nil || errors.IsInvalid(err) || errors.IsBadRequest(err) {
-		return nil
-	}
-	return err
-}
-
-func (r *EnvironmentReconciler) setEnvironmentStatus(ctx context.Context, env *platformv1alpha1.Environment, phase platformv1alpha1.EnvironmentPhase, podName, sandboxdEndpoint, reason, message string) error {
-	return r.updateEnvironmentStatus(ctx, env, func(current *platformv1alpha1.Environment) {
-		applyEnvironmentStatus(current, phase, podName, sandboxdEndpoint, reason, message, env.Status.LastActiveAt)
-		clearChildOwnershipCollision(current)
-	})
-}
-
-func (r *EnvironmentReconciler) updateEnvironmentStatus(ctx context.Context, env *platformv1alpha1.Environment, mutate func(*platformv1alpha1.Environment)) error {
-	key := client.ObjectKeyFromObject(env)
-	expectedUID := env.UID
-	expectedGeneration := env.Generation
-	var updated platformv1alpha1.Environment
-	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		if err := r.apiReader().Get(ctx, key, &updated); err != nil {
-			return err
-		}
-		if updated.UID != expectedUID {
-			return errEnvironmentIncarnationChanged
-		}
-		if updated.Generation != expectedGeneration {
-			return nil
-		}
-		before := updated.DeepCopy()
-		mutate(&updated)
-		if apiequality.Semantic.DeepEqual(before.Status, updated.Status) {
-			return nil
-		}
-		return r.Status().Update(ctx, &updated)
-	})
-	if err == nil {
-		env.Status = updated.Status
-	}
-	return err
-}
-
 func (r *EnvironmentReconciler) apiReader() client.Reader {
 	if r.APIReader != nil {
 		return r.APIReader
@@ -1921,94 +1214,3 @@ func envLabels(env *platformv1alpha1.Environment) map[string]string {
 }
 
 func ptr[T any](v T) *T { return &v }
-
-func environmentTemplateRefIndex(object client.Object) []string {
-	environment := object.(*platformv1alpha1.Environment)
-	if environment.Spec.TemplateRef == "" {
-		return nil
-	}
-	return []string{environment.Spec.TemplateRef}
-}
-
-func environmentProjectRefIndex(object client.Object) []string {
-	environment := object.(*platformv1alpha1.Environment)
-	if environment.Spec.ProjectRef == "" {
-		return nil
-	}
-	return []string{environment.Spec.ProjectRef}
-}
-
-func templateRuntimeClassIndex(object client.Object) []string {
-	template := object.(*platformv1alpha1.EnvironmentTemplate)
-	if template.Spec.RuntimeClass == "" {
-		return nil
-	}
-	return []string{template.Spec.RuntimeClass}
-}
-
-func (r *EnvironmentReconciler) environmentReferenceRequests(ctx context.Context, namespace, field, name string) []reconcile.Request {
-	var environments platformv1alpha1.EnvironmentList
-	if err := r.List(ctx, &environments, client.InNamespace(namespace), client.MatchingFields{field: name}); err != nil {
-		log.FromContext(ctx).Error(err, "list environments for reference", "field", field, "name", name)
-		return nil
-	}
-	requests := make([]reconcile.Request, 0, len(environments.Items))
-	for i := range environments.Items {
-		requests = append(requests, reconcile.Request{NamespacedName: client.ObjectKeyFromObject(&environments.Items[i])})
-	}
-	return requests
-}
-
-func (r *EnvironmentReconciler) runtimeClassReferenceRequests(ctx context.Context, name string) []reconcile.Request {
-	var templates platformv1alpha1.EnvironmentTemplateList
-	if err := r.List(ctx, &templates, client.MatchingFields{runtimeClassField: name}); err != nil {
-		log.FromContext(ctx).Error(err, "list templates for RuntimeClass", "runtimeClass", name)
-		return nil
-	}
-	var requests []reconcile.Request
-	for i := range templates.Items {
-		template := &templates.Items[i]
-		requests = append(requests, r.environmentReferenceRequests(ctx, template.Namespace, templateRefField, template.Name)...)
-	}
-	return requests
-}
-
-// SetupWithManager registers the controller with the manager.
-func (r *EnvironmentReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	r.APIReader = mgr.GetAPIReader()
-	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &platformv1alpha1.Environment{}, templateRefField, environmentTemplateRefIndex); err != nil {
-		return fmt.Errorf("index environments by template: %w", err)
-	}
-	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &platformv1alpha1.Environment{}, projectRefField, environmentProjectRefIndex); err != nil {
-		return fmt.Errorf("index environments by project: %w", err)
-	}
-	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &platformv1alpha1.EnvironmentTemplate{}, runtimeClassField, templateRuntimeClassIndex); err != nil {
-		return fmt.Errorf("index templates by RuntimeClass: %w", err)
-	}
-	return ctrl.NewControllerManagedBy(mgr).
-		For(&platformv1alpha1.Environment{}, builder.WithPredicates(predicate.Funcs{UpdateFunc: func(e event.UpdateEvent) bool {
-			old, ok1 := e.ObjectOld.(*platformv1alpha1.Environment)
-			new, ok2 := e.ObjectNew.(*platformv1alpha1.Environment)
-			return !ok1 || !ok2 || observationRelevantEnvironmentUpdate(old, new)
-		}})).
-		Owns(&corev1.Pod{}).
-		Owns(&corev1.PersistentVolumeClaim{}).
-		Owns(&networkingv1.NetworkPolicy{}).
-		Watches(&platformv1alpha1.Run{}, handler.EnqueueRequestsFromMapFunc(func(_ context.Context, object client.Object) []reconcile.Request {
-			run := object.(*platformv1alpha1.Run)
-			if run.Status.EnvironmentRef == nil {
-				return nil
-			}
-			return []reconcile.Request{{NamespacedName: types.NamespacedName{Namespace: run.Namespace, Name: run.Status.EnvironmentRef.Name}}}
-		})).
-		Watches(&platformv1alpha1.EnvironmentTemplate{}, handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, object client.Object) []reconcile.Request {
-			return r.environmentReferenceRequests(ctx, object.GetNamespace(), templateRefField, object.GetName())
-		})).
-		Watches(&platformv1alpha1.Project{}, handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, object client.Object) []reconcile.Request {
-			return r.environmentReferenceRequests(ctx, object.GetNamespace(), projectRefField, object.GetName())
-		})).
-		Watches(&nodev1.RuntimeClass{}, handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, object client.Object) []reconcile.Request {
-			return r.runtimeClassReferenceRequests(ctx, object.GetName())
-		})).
-		Complete(r)
-}
